Move all logic to main.rb
